+++ README.md
@@ -69,7 +69,7 @@ pip install rebuff
 
 ### Detect prompt injection on user input
 
-For vector database, Rebuff supports both Pinecone (default) and Chroma. To use Chroma, install Rebuff with extras: `pip install rebuff[chromadb]`
+For vector database, Rebuff supports both Pinecone (default) and Chroma. 
 
 #### With Pinecone vector database
