+++ python-sdk/tests/test_sdk.py
@@ -4,18 +4,6 @@ from rebuff.sdk import RebuffSdk, VectorDB
 from .utils import get_environment_variable
 
 
-@pytest.fixture()
-def rebuff(request) -> RebuffSdk:
-    rb = RebuffSdk(
-        get_environment_variable("OPENAI_API_KEY"),
-        request.param,
-        get_environment_variable("PINECONE_API_KEY"),
-        get_environment_variable("PINECONE_INDEX_NAME"),
-    )
-    return rb
-
-
-@pytest.fixture()
 def add_documents_to_chroma(rebuff):
     rebuff.initialize_vector_store()
     rebuff.vector_store.add_texts(
@@ -25,6 +13,20 @@ def add_documents_to_chroma(rebuff):
         metadatas={},
     )
     return None
+
+
+@pytest.fixture()
+def rebuff(request) -> RebuffSdk:
+    vector_db = request.param if hasattr(request, "param") else VectorDB.PINECONE
+    rb = RebuffSdk(
+        get_environment_variable("OPENAI_API_KEY"),
+        vector_db,
+        get_environment_variable("PINECONE_API_KEY"),
+        get_environment_variable("PINECONE_INDEX_NAME"),
+    )
+    if hasattr(request, "param") and request.param == VectorDB.CHROMA:
+        add_documents_to_chroma(rb)
+    return rb
 
 
 @pytest.fixture()
@@ -65,12 +67,6 @@ def detect_injection_arguments() -> Dict:
     return detect_injection_arguments
 
 
-@pytest.mark.parametrize(
-    "rebuff",
-    [VectorDB.PINECONE, VectorDB.CHROMA],
-    ids=["pinecone", "chroma"],
-    indirect=True,
-)
 def test_add_canary_word(user_inputs: List[str], rebuff):
     for user_input in user_inputs:
         prompt = f"Tell me a joke about\n{user_input}"
@@ -78,12 +74,6 @@ def test_add_canary_word(user_inputs: List[str], rebuff):
         assert canary_word in buffed_prompt
 
 
-@pytest.mark.parametrize(
-    "rebuff",
-    [VectorDB.PINECONE, VectorDB.CHROMA],
-    ids=["pinecone", "chroma"],
-    indirect=True,
-)
 @pytest.mark.parametrize(
     "canary_word_leaked",
     [True, False],
@@ -114,12 +104,6 @@ def test_is_canary_word_leaked(
             assert not leak_detected
 
 
-@pytest.mark.parametrize(
-    "rebuff",
-    [VectorDB.PINECONE, VectorDB.CHROMA],
-    ids=["pinecone", "chroma"],
-    indirect=True,
-)
 def test_detect_injection_heuristics(
     rebuff: RebuffSdk,
     prompt_injection_inputs: List[str],
@@ -155,7 +139,6 @@ def test_detect_injection_heuristics(
 )
 def test_detect_injection_vectorbase(
     rebuff: RebuffSdk,
-    add_documents_to_chroma,
     prompt_injection_inputs: List[str],
     benign_inputs: List[str],
     detect_injection_arguments: Dict,
@@ -182,12 +165,6 @@ def test_detect_injection_vectorbase(
         assert not rebuff_response.injection_detected
 
 
-@pytest.mark.parametrize(
-    "rebuff",
-    [VectorDB.PINECONE, VectorDB.CHROMA],
-    ids=["pinecone", "chroma"],
-    indirect=True,
-)
 def test_detect_injection_llm(
     rebuff: RebuffSdk,
     prompt_injection_inputs: List[str],
